% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_operations.R
\name{get_numids}
\alias{get_numids}
\title{Retrieves numeric IDs given alphanumeric IDs}
\usage{
get_numids(graph_obj, alpha_ids)
}
\arguments{
\item{graph_obj}{A \code{quickdag} object output by \code{\link[=qd_dag]{qd_dag()}} or \code{\link[=qd_swig]{qd_swig()}}.}

\item{alpha_ids}{A vector of alphanumeric node IDs upon which to operate.}
}
\description{
Retrieves numeric IDs given alphanumeric IDs
}
