% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_operations.R
\name{qd_set_node_attrs}
\alias{qd_set_node_attrs}
\alias{qd_set_edge_attrs}
\alias{select_nodes_by_alpha_id}
\alias{qd_select_nodes}
\alias{select_edges_by_node_alpha_id}
\alias{qd_select_edges}
\title{Graph operations}
\usage{
qd_set_node_attrs(graph_obj, alpha_ids, ..., set_op = "union")

qd_set_edge_attrs(
  graph_obj,
  ...,
  from_alpha = NULL,
  to_alpha = NULL,
  set_op = "union"
)

select_nodes_by_alpha_id(graph_obj, alpha_ids, set_op = "union")

qd_select_nodes(graph_obj, alpha_ids, set_op = "union")

select_edges_by_node_alpha_id(graph_obj, alpha_ids, set_op = "union")

qd_select_edges(graph_obj, alpha_ids, set_op = "union")
}
\arguments{
\item{graph_obj}{A \code{quickdag} object output by \code{\link[=qd_dag]{qd_dag()}} or \code{\link[=qd_swig]{qd_swig()}}.}

\item{alpha_ids}{A vector of alphanumeric node IDs upon which to operate.}

\item{...}{Pass other arguments to \code{DiagrammeR} functions:
\code{\link[DiagrammeR:set_node_attrs]{DiagrammeR::set_node_attrs()}}, \code{\link[DiagrammeR:set_edge_attrs]{DiagrammeR::set_edge_attrs()}}.}

\item{set_op}{Passed to the \code{set_op} argument of \code{\link[DiagrammeR:select_nodes_by_id]{DiagrammeR::select_nodes_by_id()}}
or \code{\link[DiagrammeR:select_edges_by_node_id]{DiagrammeR::select_edges_by_node_id()}}. Defaults to "union".}

\item{from_alpha}{A vector of alphanumeric source node IDs.}

\item{to_alpha}{A vector of alphanumeric destination node IDs.}
}
\description{
Graph operations
}
