% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_nodes}
\alias{parse_nodes}
\title{Parse nodes}
\usage{
parse_nodes(edgelist)
}
\arguments{
\item{edgelist}{A vector of edge relationships. Must be strictly organized (see
example for format).}
}
\description{
Parse nodes
}
