% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_edges}
\alias{parse_edges}
\title{Parse edges based on alpha IDs}
\usage{
parse_edges(edgestring)
}
\arguments{
\item{edgestring}{A string specifying node-edge relationships}
}
\description{
Parse edges based on alpha IDs
}
