% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qd_todagitty.R
\name{qd_adjustment_sets}
\alias{qd_adjustment_sets}
\alias{qd_todagitty}
\title{Identify variables for adjustment}
\usage{
qd_adjustment_sets(
  edgelist,
  diagram_type = "dag",
  showplot = FALSE,
  exposure,
  outcome,
  ...
)

qd_todagitty(
  edgelist,
  diagram_type = "dag",
  showplot = FALSE,
  exposure,
  outcome,
  ...
)
}
\arguments{
\item{edgelist}{A vector of edge relationships.}

\item{diagram_type}{Character identifying the diagram type. Defaults to "dag", but
user can specify another graph type (see \code{\link[dagitty:adjustmentSets]{dagitty::adjustmentSets()}}).}

\item{showplot}{Logical indicating whether to produce a dagitty plot. Defaults to
\code{FALSE}.}

\item{exposure}{Character. Specify exposure of interest. (Required)}

\item{outcome}{Character. Specifiy outcome of interest. (Required)}

\item{...}{Pass arguments to \code{\link[dagitty:adjustmentSets]{dagitty::adjustmentSets()}}.}
}
\description{
Format an edgelist and send it to dagitty to identify variable adjustment sets.
}
\details{
The \code{exposure} and \code{outcome} options map to dagitty parameters of the same name.

\code{\link[=qd_todagitty]{qd_todagitty()}} remains an alias for \code{\link[=qd_adjustment_sets]{qd_adjustment_sets()}} to avoid breaking
existing scripts, as it was the original name of this function.
}
\examples{
# feed an edgelist to qd_adjustment_sets()
edges <- c("A -> { B C D }",
           "B -> C",
           "E -> { B C }",
           "Y <- L -> A")
qd_adjustment_sets(edges, exposure = "A", outcome = "C")
qd_adjustment_sets(edges, exposure = "A", outcome = "C", type = "minimal")

# if you've already created a qd_dag() object
dag <- qd_dag(edges)
qd_adjustment_sets(dag$qd_edgelist)
}
