% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qd_save.R
\name{qd_save}
\alias{qd_save}
\title{Save graph objects as images}
\usage{
qd_save(graph, file_name, ..., embed = FALSE, kg = NULL)
}
\arguments{
\item{graph}{A graph object produced by [qd_dag()]}

\item{file_name}{A filename (possibly including a path) to which to save the image.
File type will be determined by the extension provided (pdf, svg, png, ps). See
[DiagrammeR::export_graph()] for details.}

\item{...}{Pass other arguments to [DiagrammeR::render_graph()].}

\item{embed}{Defaults to `FALSE`. Automatically set to `TRUE` by [qd_embed()].}

\item{kg}{A list allowing the user to set arguments for [knitr::graphics()] (except for
`path`).}
}
\description{
[qd_save()] is a light wrapper around [DiagrammeR::export_graph()], which
enables exporting high-quality and/or scalable graphics for both print and online.
For convenience in computational notebooks like quarto or RMarkdown documents,
[qd_save()] also allows the user to embed the image via [knitr::include_graphics()].
}
\examples{
# redirect file to temporary directory (example only)
file <- file.path(tempdir(), "dag.pdf")

# save without embedding
dag <- qd_dag(c("L -> {A Y}", "A -> Y"))
qd_save(dag, file)

# embed from code chunk using knitr::include_graphics()
qd_save(dag, file, embed = TRUE)

# `title` passed to DiagrammeR::export_graph()
qd_save(dag, file, title = "Demo")

# clean up temporary directory
file.remove(file)
}
