% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{quickDAG_setup}
\alias{quickDAG_setup}
\title{Package Setup}
\usage{
quickDAG_setup()
}
\description{
A temporary workaround until devtools::install_github gets a bugfix that properly installs all package dependencies (and sub-dependencies).
}
