% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_edgestring}
\alias{parse_edgestring}
\alias{parse_edges}
\title{Parse edges in a graph specification}
\usage{
parse_edgestring(edgestring)

parse_edges(edgelist)
}
\arguments{
\item{edgestring}{A single string containing some set of node-edge relationships}

\item{edgelist}{A vector of edge relationships. Must be strictly organized (see
example for format).}
}
\description{
Parse edges in a graph specification
}
