% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qd_dag.R
\name{qd_dag}
\alias{qd_dag}
\title{Generate a graph object}
\usage{
qd_dag(
  edgelist,
  node_labs = NULL,
  node_aes_opts = list(),
  edge_aes_opts = list(),
  verbose = FALSE,
  check_dag = TRUE,
  theme = "base",
  ...
)
}
\arguments{
\item{edgelist}{A vector of node-edge relationships.}

\item{node_labs}{A named character vector containing label names. Defaults
to \code{NULL}.}

\item{node_aes_opts}{A list feeding aesthetic options for nodes to
\code{\link[DiagrammeR:node_aes]{DiagrammeR::node_aes()}}. Defaults to empty list.}

\item{edge_aes_opts}{A list feeding aesthetic options for edges to
\code{\link[DiagrammeR:edge_aes]{DiagrammeR::edge_aes()}}. Defaults to empty list.}

\item{verbose}{Indicate whether to print node and edge dataframes to the console.
See Details below. Defaults to \code{TRUE}.}

\item{check_dag}{Logical. Check whether the graph conforms to the rules of DAGs.
Defaults to \code{TRUE}.}

\item{theme}{Choose theme for plot output. Defaults to \code{"base"}. Setting theme to
\code{NULL} will use DiagrammeR's \code{NULL} attribute theme.}

\item{...}{Pass optional \code{conditioned} argument to \code{\link[=qd_themes]{qd_themes()}}.}
}
\description{
Provide simple syntax specifying paths between nodes to generate a graph object.
}
\details{
Leaving the \code{verbose} set to \code{TRUE} may be advisable to ensure labels and IDs have
not been mismatched. By default, \code{\link[=qd_dag]{qd_dag()}} alphabetizes nodes included in \code{edgelist}
and does the same for \code{node_labs} under a first assumption that labels will begin with
the same letter as their corresponding \code{alpha_id}, which may not always be the case.
}
\examples{
# Provide a list of edges, with nodes specified as letters or single words.
edges <- c("A -> { B C } <- L",
           "B -> C")

# make a DAG object and render the graph using the default theme
g.obj <- qd_dag(edges)
DiagrammeR::render_graph(g.obj)

# Pass labels and aesthetic options for nodes or edges
g.obj2 <- qd_dag(edges,
                 node_labs = c("A" = "Alcohol",
                               "B" = "BP",
                               "C" = "CVD",
                               "L" = "State"),
                 node_aes_opts = list(shape = "plaintext",
                                      fillcolor = "none",
                                      color = "black"),
                 edge_aes_opts = list(arrowsize = 0.5,
                                      color = "gray"),
                 theme = NULL)
DiagrammeR::render_graph(g.obj2)

}
