% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qd_save.R
\name{qd_embed}
\alias{qd_embed}
\title{Embed diagrams in RMarkdown}
\usage{
qd_embed(...)
}
\arguments{
\item{...}{Pass arguments to [qd_save()].}
}
\description{
A wrapper around [qd_save()] meant for use within R code chunks in
  RMarkdown documents.
}
