% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qd_save.R
\name{qd_save}
\alias{qd_save}
\title{Save graph objects}
\usage{
qd_save(graph, filename = NULL, filetype = "pdf", embed = F, ...)
}
\arguments{
\item{graph}{Either a DiagrammeR graph object or a diagram generated by \code{render_graph()}.}

\item{filename}{String for filename. Defaults to \code{NULL}.}

\item{filetype}{One of \code{pdf}, \code{eps}, \code{pdf}, \code{svg}. Defaults to \code{pdf}.}

\item{embed}{Defaults to FALSE. Automatically set to TRUE by \code{qd_embed()}.}

\item{...}{Pass arguments to \code{render_graph()}, e.g., width and height (pixels) for .png files.}
}
\description{
Export high-quality, scalable graphics for both print and online.
}
\section{Functions}{
\itemize{
\item \code{qd_save()}: 

}}
