% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qd_swig.R
\name{qd_swig}
\alias{qd_swig}
\title{Generate a single word intervention graph (SWIG) template}
\usage{
qd_swig(graph.obj, fixed.nodes)
}
\arguments{
\item{graph.obj}{A DAG object created by \code{qd_dag()}.}

\item{fixed.nodes}{A vector containing the nodes to be intervened upon.

Suggestions and bug reports welcome at \url{https://github.com/jrgant/quickDAG/issues}.


Packages used: DiagrammeR, stringr, purrr}
}
\description{
Provide simple syntax specifying paths between nodes to generate a graph object.
}
\examples{
# Provide a DAG object and a list of nodes to be fixed
qd_swig(dag, c("A", "L")) \%>\% render_graph()

}
