% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qd_swig.R
\name{qd_swig}
\alias{qd_swig}
\title{Generate a single-world intervention graph (SWIG)}
\usage{
qd_swig(
  graph.obj,
  fixed.nodes,
  custom.values = NULL,
  fixed.sep = "vlin",
  sep.point.size = 15
)
}
\arguments{
\item{graph.obj}{A DAG object created by \code{qd_dag()}.}

\item{fixed.nodes}{A vector containing the nodes to be intervened upon.}

\item{custom.values}{A named vector containing alternative labels identifying explicit values for fixed nodes (e.g., \code{c("A" = "1")}).}

\item{fixed.sep}{A character string indicating which character to use as a separator in fixed nodes. Defaults to \code{"vlin"}. Run \code{sep_opts(TRUE)} for available options.}

\item{sep.point.size}{A numerical value specifying the point size for fixed node separators.}
}
\description{
Take a DAG graph object and, in the simplest case, create a single-world intervention template corresponding
to a world in which the fixed nodes are set to a given value. Alternatively, tell qd_swig which values fixed nodes
will be set to.
}
\examples{
# Provide a DAG object and a list of nodes to be fixed
library(magrittr)
edges <- c("A -> Y",
           "L -> { A Y }")

dag  <- qd_dag(edges)

swig <- dag \%>\%
        qd_swig(fixed.nodes = "A",
                custom.values = c("A" = "1"))

swig \%>\% DiagrammeR::render_graph()

}
