% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{qd_themes}
\alias{qd_themes}
\alias{theme_qd_base}
\alias{theme_qd_circles}
\alias{theme_qd_pearl}
\alias{get_conditioned_nodes}
\alias{cleanup_existing_theme}
\title{Diagram themes}
\usage{
qd_themes(graph_obj, theme, ...)

theme_qd_base(graph_obj, font = "serif", ...)

theme_qd_circles(graph_obj, font = "serif", ...)

theme_qd_pearl(
  graph_obj,
  font = "serif",
  pointsize = getOption("quickdag.pearl_pointsize"),
  linewidths = getOption("quickdag.pearl_linewidths"),
  fontsize = getOption("quickdag.pearl_fontsize"),
  fontcolor = getOption("quickdag.pearl_fontcolor"),
  ...
)

get_conditioned_nodes(graph_obj, conditioned = NULL)

cleanup_existing_theme(graph_obj)
}
\arguments{
\item{graph_obj}{A DAG object created by \code{\link[=qd_dag]{qd_dag()}}.}

\item{theme}{A character string indicating the theme to use. Defaults to "base".
Set to \code{NULL} to use GraphViz defaults.}

\item{...}{Pass arguments to theme call (e.g., \code{\link[=theme_qd_base]{theme_qd_base()}}), such as
\code{conditioned} or \code{font}.}

\item{font}{A character vector indicating the font family to use for node labels.
Defaults to "serif".}

\item{conditioned}{A character vector indicating which nodes are conditioned upon.
The shape for these nodes will be set to "rectangle".}
}
\description{
Apply various pre-fabricated themes to diagrams.
}
