% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{qd_themes}
\alias{qd_themes}
\alias{theme_base}
\alias{theme_circles}
\alias{theme_dots}
\alias{get_conditioned_nodes}
\title{Diagram themes}
\usage{
qd_themes(graph_obj, theme, ...)

theme_base(graph_obj, font = "serif", ...)

theme_circles(graph_obj, font = "serif", ...)

theme_dots(graph_obj, font = "serif", ...)

get_conditioned_nodes(graph_obj, conditioned = NULL)
}
\arguments{
\item{graph_obj}{A DAG object created by [qd_dag()].}

\item{theme}{A character string indicating the theme to use. Defaults to "base".
Set to `NULL` to use GraphViz defaults.}

\item{...}{Pass arguments to theme call (e.g., [theme_base()]), such as `conditioned`
or `font`.}

\item{font}{A character vector indicating the font family to use for node labels.
Defaults to "serif".}

\item{conditioned}{A character vector indicating which nodes are conditioned upon.
The shape for these nodes will be set to "square".}
}
\description{
Apply various pre-fabricated themes to diagrams.
}
