% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickdag-package.R
\docType{package}
\name{quickdag-package}
\alias{quickdag}
\alias{quickdag-package}
\title{quickdag: Make Beautiful Directed Acyclic Graphs}
\description{
This package allows the user the save directed acyclic graphs (DAGs) generated in DiagrammeR and export them to pdf, png, or svg format. Users may optionally choose to view the DAG without saving or to embed it in an RMarkdown document. In addition, quickDAG allows the user easily to reformat DAGs as single-world intervention graph (SWIG) templates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jrgant/quickDAG}
  \item Report bugs at \url{https://github.com/jrgant/quickDAG/issues}
}

}
\author{
\strong{Maintainer}: Jason Gantenberg \email{jason.gantenberg@gmail.com} (\href{https://orcid.org/0000-0002-8144-4273}{ORCID})

}
\keyword{internal}
