% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep_opts.R
\name{sep_opts}
\alias{sep_opts}
\title{View options for fixed node separator characters}
\usage{
sep_opts(table = FALSE)
}
\arguments{
\item{table}{Logical to show or hide HTML table display of available characters.
Defaults to \code{FALSE}.}
}
\description{
Preview character options for use as the fixed node separator in SWIGs.
}
